+++ NEWS.md
@@ -41,7 +41,7 @@
 * in `describe` the specification of custom aggregation is now `function => name`;
   old `name => function` order is now deprecated
   ([#2401](https://github.com/JuliaData/DataFrames.jl/pull/2401))
-* `unstack` now does produce rows and columns in their order of appereance in the source
+* `unstack` now does produce rows and columns in the order which `groupby` produces
    and has two new keyword arguments `allowmissing` and `allowduplicates`
   ([#2494](https://github.com/JuliaData/DataFrames.jl/pull/2494))